+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add `getPageScrollElm` … returns `<html/>` unless `<body/>` is
+  scrollable
 
 ## 4.7.3